call get_articles from one way
