+++ tasks/maintenance/read.py
@@ -6,15 +6,14 @@ import traceback
 
 
 def main(*args: str) -> int:
+    # todo: mereg with read.py in webcite task
     try:
         thread_number = 1
         time_before_start = int(sys.argv[1])
-
         if time_before_start == 2540:
             thread_number = 3
         elif time_before_start == 500:
             thread_number = 2
-
         pages = get_pages(time_before_start)
         conn, cursor = create_database_table(maintenance_db_name)
         save_pages_to_db(pages, conn, cursor, thread_number=thread_number)
