+++ tasks/webcite/module.py
@@ -35,46 +35,6 @@ FROM (
 
 
 
-def get_articles(cursor, thread_number):
-    if thread_number == 1:
-        thread_number = 0
-    query1 = """
-        SELECT id, title, thread
-        FROM pages
-        WHERE status = 0 AND thread = 1
-        ORDER BY date ASC
-        LIMIT 200 OFFSET ?;
-    """
-
-    query2 = """
-        SELECT id, title, thread
-        FROM pages
-        WHERE status = 0 AND thread = 2
-        ORDER BY date ASC
-        LIMIT 150 OFFSET ?;
-    """
-
-    query3 = """
-        SELECT id, title, thread
-        FROM pages
-        WHERE status = 0 AND thread = 3
-        ORDER BY date ASC
-        LIMIT 100 OFFSET ?;
-    """
-
-    cursor.execute(query1, (thread_number,))
-    result1 = cursor.fetchall()
-
-    cursor.execute(query2, (thread_number,))
-    result2 = cursor.fetchall()
-
-    cursor.execute(query3, (thread_number,))
-    result3 = cursor.fetchall()
-
-    rows = result1 + result2 + result3
-    return rows
-
-
 def check_status():
     site = pywikibot.Site()
     title = "مستخدم:LokasBot/الإبلاغ عن رابط معطوب أو مؤرشف"
