+++ tasks/webcite/read.py
@@ -1,19 +1,17 @@
 import sys
-
 from core.utils.sqlite import create_database_table, webcite_db_name, save_pages_to_db
 from tasks.webcite.module import get_pages
 
 
 def main(*args: str) -> int:
     try:
+        # todo: mereg with read.py in maintenance task
         thread_number = 1
         time_before_start = int(sys.argv[1])
-
         if time_before_start == 2540:
             thread_number = 3
         elif time_before_start == 500:
             thread_number = 2
-
         pages = get_pages(time_before_start)
         conn, cursor = create_database_table(webcite_db_name)
         save_pages_to_db(pages, conn, cursor, thread_number=thread_number)
